Rearranged directory structure
