+++ .gitignore
@@ -30,5 +30,4 @@
 *.exe
 *.out
 *.app
-Flood_test
-Mat_test
+bin/*
