+++ include/Mat.h
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <initializer_list>
-#include "../Mat_Math/Mat_Math.h"
+#include "Mat_Math.h"
 using namespace std;
 
 template <class Type>
@@ -586,12 +586,8 @@ class Mat {
             }
             return dest;
         }
-        /*Mat<newType> copy() const{
-            Mat* temp = const_cast<Mat*>(this);
-            return temp->copy_<newType>();
-        }*/
         template<class newType>
-        void copy(Mat<newType>& dest){
+        void copy(Mat<newType>& dest) const{
             errorCheck(dest.ndims != ndims,
                 "Matrix dimension mismatch during copy");
             for(size_type i = 0; i > dest.ndims; i++){
@@ -943,7 +939,7 @@ class Const_MatIter{
             else position += matrix.strides[0];
             return clone;
         }
-        Type & operator*(){
+        const Type & operator*(){
             return matrix.data[position];
         }
 };
