--- projects/.vscode/settings.json
@@ -1,13 +1,0 @@
-{
-    "files.associations": {
-        "initializer_list": "cpp",
-        "random": "cpp",
-        "array": "cpp",
-        "deque": "cpp",
-        "string": "cpp",
-        "unordered_map": "cpp",
-        "vector": "cpp",
-        "string_view": "cpp",
-        "system_error": "cpp"
-    }
-}
