--- projects/.vscode/tasks.json
@@ -1,59 +1,0 @@
-{
-    "version": "2.0.0",
-    "tasks": [
-        {
-            "type": "shell",
-            "label": "C/C++: g++ build active file",
-            "command": "/usr/bin/g++",
-            "args": [
-                "-g",
-                "${file}",
-                "-o",
-                "${fileDirname}/${fileBasenameNoExtension}",
-                "-Wall",
-                "-Werror",
-                "--pedantic",
-                "-O0"
-            ],
-            "options": {
-                "cwd": "/usr/bin"
-            },
-            "problemMatcher": [
-                "$gcc"
-            ],
-            "group": {
-                "kind": "build",
-                "isDefault": true
-            }
-        },
-        {
-            "type": "shell",
-            "label": "g++ build active file",
-            "command": "/usr/bin/g++",
-            "args": [
-                "-g",
-                "${file}",
-                "-o",
-                "${fileDirname}/${fileBasenameNoExtension}",
-                "-Wall",
-                "-Werror",
-                "--pedantic",
-                "-O0"
-            ],
-            "options": {
-                "cwd": "/usr/bin"
-            }
-        },
-        {
-            "type": "shell",
-            "label": "Run through Valgrind",
-            "command": "valgrind",
-            "args": [
-                "-s",
-                "--leak-check=full",
-                "${fileDirname}/${fileBasenameNoExtension}"
-            ],
-            "problemMatcher": []
-        }
-    ]
-}
